Adding notifications actions

+++ components/questions/QuestionItemActions.tsx
@@ -2,6 +2,7 @@
 // import Image from "next/image";
 import { deleteQuestion } from "@/library/actions/questions.actions";
 import { useSession } from "next-auth/react";
+import { ToastContainer, toast } from "react-toastify";
 import { usePathname } from "next/navigation";
 
 const QuestionItemActions = ({ id }: { id: string }) => {
@@ -13,6 +14,11 @@ const QuestionItemActions = ({ id }: { id: string }) => {
       "Are you sure you want to delete this question?"
     );
 
+    if (!session) {
+      toast.error("You must be logged in to delete a question.");
+      return;
+    }
+
     if (confirm) {
       // @ts-ignore
       await deleteQuestion(id, session?.user.id, pathname);
@@ -21,6 +27,7 @@ const QuestionItemActions = ({ id }: { id: string }) => {
 
   return (
     <div>
+      <ToastContainer />
       <div className="flex gap-4 items-center justify-center flex-wrap">
         <div>
           <button
